requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.3
    # via
    #   discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
discord-py==2.3.2
    # via -r requirements.in
disputils==0.1.1
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
grpcio==1.62.0
    # via
    #   -r requirements.in
    #   grpcio-reflection
grpcio-reflection==1.62.0
    # via -r requirements.in
idna==3.6
    # via
    #   requests
    #   yarl
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
numpy==1.24.4
    # via
    #   pandas
    #   scipy
pandas==2.0.3
    # via -r requirements.in
protobuf==4.25.3
    # via grpcio-reflection
python-dateutil==2.8.2
    # via pandas
pytz==2024.1
    # via pandas
requests==2.31.0
    # via -r requirements.in
scipy==1.10.1
    # via -r requirements.in
six==1.16.0
    # via
    #   python-dateutil
    #   trueskill
trueskill==0.4.5
    # via -r requirements.in
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
yarl==1.9.4
    # via aiohttp
